Перенёс методы конвертации цвета в строку и обратно в extension для UIColor.
